Helper base class for non-blittable element marshalling
