+++ src/libraries/System.Runtime.InteropServices/gen/Microsoft.Interop.SourceGeneration/Marshalling/StatefulMarshallingStrategy.cs
@@ -543,7 +543,7 @@ namespace Microsoft.Interop
     /// <summary>
     /// Marshaller that enables support for marshalling non-blittable elements of a collection via a native type that implements the LinearCollection marshalling spec.
     /// </summary>
-    internal sealed class StatefulLinearCollectionNonBlittableElementsMarshalling : ICustomTypeMarshallingStrategy
+    internal sealed class StatefulLinearCollectionNonBlittableElementsMarshalling : NonBlittableElementsMarshalling, ICustomTypeMarshallingStrategy
     {
         private readonly ICustomTypeMarshallingStrategy _innerMarshaller;
         private readonly MarshallerShape _shape;
@@ -559,6 +559,7 @@ namespace Microsoft.Interop
             IMarshallingGenerator elementMarshaller,
             TypePositionInfo elementInfo,
             ExpressionSyntax numElementsExpression)
+            : base (unmanagedElementType, elementMarshaller, elementInfo)
         {
             _innerMarshaller = innerMarshaller;
             _shape = shape;
@@ -584,51 +585,14 @@ namespace Microsoft.Interop
 
             if (!info.IsByRef && info.ByValueContentsMarshalKind == ByValueContentsMarshalKind.Out)
             {
-                // If the parameter is marshalled by-value [Out], then we don't marshal the contents of the collection.
-                // We do clear the span, so that if the invoke target doesn't fill it, we aren't left with undefined content.
-                // <marshaller>.GetUnmanagedValuesDestination().Clear();
-                yield return ExpressionStatement(
-                    InvocationExpression(
-                        MemberAccessExpression(
-                            SyntaxKind.SimpleMemberAccessExpression,
-                            GetUnmanagedValuesDestination(info, context),
-                            IdentifierName("Clear"))));
+                yield return GenerateByValueOutMarshalStatement(info, context);
                 yield break;
             }
-
-            string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-            string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
 
             // ReadOnlySpan<T> <managedSpan> = <marshaller>.GetManagedValuesSource()
             // Span<TUnmanagedElement> <nativeSpan> = <marshaller>.GetUnmanagedValuesDestination()
             // << marshal contents >>
-            yield return Block(
-                LocalDeclarationStatement(VariableDeclaration(
-                    GenericName(
-                        Identifier(TypeNames.System_ReadOnlySpan),
-                        TypeArgumentList(SingletonSeparatedList(_elementInfo.ManagedType.Syntax))),
-                    SingletonSeparatedList(
-                        VariableDeclarator(Identifier(managedSpanIdentifier))
-                        .WithInitializer(EqualsValueClause(
-                            GetManagedValuesSource(info, context)))))),
-                LocalDeclarationStatement(VariableDeclaration(
-                    GenericName(
-                        Identifier(TypeNames.System_Span),
-                        TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                    SingletonSeparatedList(
-                        VariableDeclarator(
-                            Identifier(nativeSpanIdentifier))
-                        .WithInitializer(EqualsValueClause(
-                            GetUnmanagedValuesDestination(info, context)))))),
-                MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                    info,
-                    context,
-                    _elementInfo,
-                    _elementMarshaller,
-                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
-                        IdentifierName(MarshallerHelpers.GetManagedSpanIdentifier(info, context)),
-                        IdentifierName("Length")),
-                    StubCodeContext.Stage.Marshal));
+            yield return GenerateMarshalStatement(info, context);
         }
 
         public IEnumerable<StatementSyntax> GenerateNotifyForSuccessfulInvokeStatements(TypePositionInfo info, StubCodeContext context) => _innerMarshaller.GenerateNotifyForSuccessfulInvokeStatements(info, context);
@@ -641,11 +605,22 @@ namespace Microsoft.Interop
             if (!_shape.HasFlag(MarshallerShape.ToManaged))
                 yield break;
 
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
 
             if (!info.IsByRef && info.ByValueContentsMarshalKind.HasFlag(ByValueContentsMarshalKind.Out))
             {
-                yield return GenerateByValueUnmarshalStatement(info, context);
+                // int <numElements> = <GetManagedValuesSource>.Length;
+                yield return LocalDeclarationStatement(
+                    VariableDeclaration(
+                        PredefinedType(Token(SyntaxKind.IntKeyword)),
+                        SingletonSeparatedList(
+                            VariableDeclarator(numElementsIdentifier)
+                                .WithInitializer(EqualsValueClause(
+                                    MemberAccessExpression(
+                                        SyntaxKind.SimpleMemberAccessExpression,
+                                        GetManagedValuesSource(info, context),
+                                        IdentifierName("Length")))))));
+                yield return GenerateByValueOutUnmarshalStatement(info, context);
             }
             else
             {
@@ -657,39 +632,10 @@ namespace Microsoft.Interop
                             VariableDeclarator(numElementsIdentifier)
                                 .WithInitializer(EqualsValueClause(_numElementsExpression)))));
 
-                string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-                string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
-
                 // ReadOnlySpan<TUnmanagedElement> <nativeSpan> = <marshaller>.GetUnmanagedValuesSource(<nativeIdentifier>, <numElements>)
                 // Span<T> <managedSpan> = <marshaller>.GetManagedValuesDestination(<managedIdentifier>)
                 // << unmarshal contents >>
-                yield return Block(
-                    LocalDeclarationStatement(VariableDeclaration(
-                        GenericName(
-                            Identifier(TypeNames.System_ReadOnlySpan),
-                            TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                        SingletonSeparatedList(
-                            VariableDeclarator(
-                                Identifier(nativeSpanIdentifier))
-                            .WithInitializer(EqualsValueClause(
-                                GetUnmanagedValuesSource(info, context)))))),
-                    LocalDeclarationStatement(VariableDeclaration(
-                        GenericName(
-                            Identifier(TypeNames.System_Span),
-                            TypeArgumentList(SingletonSeparatedList(_elementInfo.ManagedType.Syntax))),
-                        SingletonSeparatedList(
-                            VariableDeclarator(
-                                Identifier(managedSpanIdentifier))
-                            .WithInitializer(EqualsValueClause(
-                                GetManagedValuesDestination(info, context)))))),
-                    MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                        info,
-                        context,
-                        _elementInfo,
-                        _elementMarshaller,
-                        IdentifierName(numElementsIdentifier),
-                        StubCodeContext.Stage.UnmarshalCapture,
-                        StubCodeContext.Stage.Unmarshal));
+                yield return GenerateUnmarshalStatement(info, context);
             }
 
             foreach (StatementSyntax statement in _innerMarshaller.GenerateUnmarshalStatements(info, context))
@@ -702,10 +648,7 @@ namespace Microsoft.Interop
 
         public bool UsesNativeIdentifier(TypePositionInfo info, StubCodeContext context) => true;
 
-        private static string GetNumElementsIdentifier(TypePositionInfo info, StubCodeContext context)
-            => context.GetAdditionalIdentifier(info, "numElements");
-
-        private static InvocationExpressionSyntax GetUnmanagedValuesDestination(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetUnmanagedValuesDestination(TypePositionInfo info, StubCodeContext context)
         {
             string marshaller = StatefulValueMarshalling.GetMarshallerIdentifier(info, context);
 
@@ -718,7 +661,7 @@ namespace Microsoft.Interop
                 ArgumentList());
         }
 
-        private static InvocationExpressionSyntax GetManagedValuesSource(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetManagedValuesSource(TypePositionInfo info, StubCodeContext context)
         {
             string marshaller = StatefulValueMarshalling.GetMarshallerIdentifier(info, context);
 
@@ -731,10 +674,10 @@ namespace Microsoft.Interop
                 ArgumentList());
         }
 
-        private static InvocationExpressionSyntax GetUnmanagedValuesSource(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetUnmanagedValuesSource(TypePositionInfo info, StubCodeContext context)
         {
             string marshaller = StatefulValueMarshalling.GetMarshallerIdentifier(info, context);
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
 
             // <marshaller>.GetUnmanagedValuesSource(<numElements>)
             return InvocationExpression(
@@ -746,10 +689,10 @@ namespace Microsoft.Interop
                     Argument(IdentifierName(numElementsIdentifier)))));
         }
 
-        private static InvocationExpressionSyntax GetManagedValuesDestination(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetManagedValuesDestination(TypePositionInfo info, StubCodeContext context)
         {
             string marshaller = StatefulValueMarshalling.GetMarshallerIdentifier(info, context);
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
 
             // <marshaller>.GetManagedValuesDestination(<numElements>)
             return InvocationExpression(
@@ -759,91 +702,6 @@ namespace Microsoft.Interop
                     IdentifierName(ShapeMemberNames.LinearCollection.Stateless.GetManagedValuesDestination)),
                 ArgumentList(SingletonSeparatedList(
                     Argument(IdentifierName(numElementsIdentifier)))));
-        }
-
-        private StatementSyntax GenerateByValueUnmarshalStatement(TypePositionInfo info, StubCodeContext context)
-        {
-            // Use ManagedSource and NativeDestination spans for by-value marshalling since we're just marshalling back the contents,
-            // not the array itself.
-            // This code is ugly since we're now enforcing readonly safety with ReadOnlySpan for all other scenarios,
-            // but this is an uncommon case so we don't want to design the API around enabling just it.
-            string numElementsIdentifier = context.GetAdditionalIdentifier(info, "numElements");
-
-            // <numElements> = <marshaller>.GetManagedValuesSource().Length;
-            ExpressionStatementSyntax numElementsStatement = ExpressionStatement(
-                AssignmentExpression(
-                    SyntaxKind.SimpleAssignmentExpression,
-                    IdentifierName(numElementsIdentifier),
-                    MemberAccessExpression(
-                        SyntaxKind.SimpleMemberAccessExpression,
-                        GetManagedValuesSource(info, context),
-                        IdentifierName("Length"))));
-
-            string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-
-            // Span<TElement> <managedSpan> = MemoryMarshal.CreateSpan(ref Unsafe.AsRef(in <marshaller>.GetManagedValuesSource().GetPinnableReference(), <numElements>));
-            LocalDeclarationStatementSyntax managedValuesDeclaration = LocalDeclarationStatement(VariableDeclaration(
-                GenericName(
-                    Identifier(TypeNames.System_Span),
-                    TypeArgumentList(
-                        SingletonSeparatedList(_elementInfo.ManagedType.Syntax))
-                ),
-                SingletonSeparatedList(VariableDeclarator(managedSpanIdentifier).WithInitializer(EqualsValueClause(
-                    InvocationExpression(
-                        MemberAccessExpression(
-                            SyntaxKind.SimpleMemberAccessExpression,
-                            ParseName(TypeNames.System_Runtime_InteropServices_MemoryMarshal),
-                            IdentifierName("CreateSpan")))
-                    .WithArgumentList(
-                        ArgumentList(
-                            SeparatedList(
-                                new[]
-                                {
-                                    Argument(
-                                        InvocationExpression(
-                                            MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
-                                                ParseName(TypeNames.System_Runtime_CompilerServices_Unsafe),
-                                                IdentifierName("AsRef")),
-                                            ArgumentList(SingletonSeparatedList(
-                                                Argument(
-                                                    InvocationExpression(
-                                                        MemberAccessExpression(
-                                                            SyntaxKind.SimpleMemberAccessExpression,
-                                                            GetManagedValuesSource(info, context),
-                                                            IdentifierName("GetPinnableReference")),
-                                                            ArgumentList()))
-                                                .WithRefKindKeyword(
-                                                    Token(SyntaxKind.InKeyword))))))
-                                    .WithRefKindKeyword(
-                                        Token(SyntaxKind.RefKeyword)),
-                                    Argument(
-                                        IdentifierName(numElementsIdentifier))
-                                }))))))));
-
-            // Span<TUnmanagedElement> <nativeSpan> = <marshaller>.GetUnmanagedValuesDestination()
-            string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
-            LocalDeclarationStatementSyntax unmanagedValuesDeclaration = LocalDeclarationStatement(VariableDeclaration(
-                GenericName(
-                    Identifier(TypeNames.System_Span),
-                    TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                SingletonSeparatedList(
-                    VariableDeclarator(
-                        Identifier(nativeSpanIdentifier))
-                    .WithInitializer(EqualsValueClause(
-                        GetUnmanagedValuesDestination(info, context))))));
-
-            return Block(
-                numElementsStatement,
-                managedValuesDeclaration,
-                unmanagedValuesDeclaration,
-                MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                    info,
-                    context,
-                    _elementInfo,
-                    _elementMarshaller,
-                    IdentifierName(numElementsIdentifier),
-                    StubCodeContext.Stage.UnmarshalCapture,
-                    StubCodeContext.Stage.Unmarshal));
         }
     }
 }
