+++ src/libraries/System.Runtime.InteropServices/gen/Microsoft.Interop.SourceGeneration/Marshalling/StatelessMarshallingStrategy.cs
@@ -526,7 +526,7 @@ namespace Microsoft.Interop
     /// <summary>
     /// Marshaller that enables support for marshalling non-blittable elements of a collection via a native type that implements the LinearCollection marshalling spec.
     /// </summary>
-    internal sealed class StatelessLinearCollectionNonBlittableElementsMarshalling : ICustomTypeMarshallingStrategy
+    internal sealed class StatelessLinearCollectionNonBlittableElementsMarshalling : NonBlittableElementsMarshalling, ICustomTypeMarshallingStrategy
     {
         private readonly TypeSyntax _marshallerTypeSyntax;
         private readonly TypeSyntax _nativeTypeSyntax;
@@ -544,6 +544,7 @@ namespace Microsoft.Interop
             IMarshallingGenerator elementMarshaller,
             TypePositionInfo elementInfo,
             ExpressionSyntax numElementsExpression)
+            : base (unmanagedElementType, elementMarshaller, elementInfo)
         {
             _marshallerTypeSyntax = marshallerTypeSyntax;
             _nativeTypeSyntax = nativeTypeSyntax;
@@ -565,7 +566,7 @@ namespace Microsoft.Interop
                 yield break;
 
             (string managedIdentifier, string nativeIdentifier) = context.GetIdentifiers(info);
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
 
             // <nativeIdentifier> = <marshallerType>.AllocateContainerForUnmanagedElements(<managedIdentifier>, out <numElements>);
             yield return ExpressionStatement(
@@ -585,51 +586,11 @@ namespace Microsoft.Interop
 
             if (!info.IsByRef && info.ByValueContentsMarshalKind == ByValueContentsMarshalKind.Out)
             {
-                // If the parameter is marshalled by-value [Out], then we don't marshal the contents of the collection.
-                // We do clear the span, so that if the invoke target doesn't fill it, we aren't left with undefined content.
-                // <marshallerType>.GetUnmanagedValuesDestination(<nativeIdentifier>, <numElements>).Clear();
-                yield return ExpressionStatement(
-                    InvocationExpression(
-                        MemberAccessExpression(
-                            SyntaxKind.SimpleMemberAccessExpression,
-                            GetUnmanagedValuesDestination(info, context),
-                            IdentifierName("Clear"))));
+                yield return GenerateByValueOutMarshalStatement(info, context);
                 yield break;
             }
 
-            string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-            string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
-
-            // ReadOnlySpan<T> <managedSpan> = <marshallerType>.GetManagedValuesSource(<managedIdentifer>)
-            // Span<TUnmanagedElement> <nativeSpan> = <marshallerType>.GetUnmanagedValuesDestination(<nativeIdentifier>, <numElements>)
-            // << marshal contents >>
-            yield return Block(
-                LocalDeclarationStatement(VariableDeclaration(
-                    GenericName(
-                        Identifier(TypeNames.System_ReadOnlySpan),
-                        TypeArgumentList(SingletonSeparatedList(_elementInfo.ManagedType.Syntax))),
-                    SingletonSeparatedList(
-                        VariableDeclarator(Identifier(managedSpanIdentifier))
-                        .WithInitializer(EqualsValueClause(
-                            GetManagedValuesSource(info, context)))))),
-                LocalDeclarationStatement(VariableDeclaration(
-                    GenericName(
-                        Identifier(TypeNames.System_Span),
-                        TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                    SingletonSeparatedList(
-                        VariableDeclarator(
-                            Identifier(nativeSpanIdentifier))
-                        .WithInitializer(EqualsValueClause(
-                            GetUnmanagedValuesDestination(info, context)))))),
-                MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                    info,
-                    context,
-                    _elementInfo,
-                    _elementMarshaller,
-                    MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
-                        IdentifierName(MarshallerHelpers.GetManagedSpanIdentifier(info, context)),
-                        IdentifierName("Length")),
-                    StubCodeContext.Stage.Marshal));
+            yield return GenerateMarshalStatement(info, context);
         }
 
         public IEnumerable<StatementSyntax> GenerateNotifyForSuccessfulInvokeStatements(TypePositionInfo info, StubCodeContext context) => Array.Empty<StatementSyntax>();
@@ -642,7 +603,7 @@ namespace Microsoft.Interop
                 VariableDeclaration(
                     PredefinedType(Token(SyntaxKind.IntKeyword)),
                     SingletonSeparatedList(
-                        VariableDeclarator(GetNumElementsIdentifier(info, context)))));
+                        VariableDeclarator(MarshallerHelpers.GetNumElementsIdentifier(info, context)))));
         }
 
         public IEnumerable<StatementSyntax> GenerateUnmarshalCaptureStatements(TypePositionInfo info, StubCodeContext context) => Array.Empty<StatementSyntax>();
@@ -653,14 +614,24 @@ namespace Microsoft.Interop
                 yield break;
 
             (string managedIdentifier, string nativeIdentifier) = context.GetIdentifiers(info);
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
 
             if (!info.IsByRef && info.ByValueContentsMarshalKind.HasFlag(ByValueContentsMarshalKind.Out))
             {
-                yield return GenerateByValueUnmarshalStatement(info, context);
+                // <numElements> = <GetManagedValuesSource.Length;
+                yield return ExpressionStatement(
+                    AssignmentExpression(
+                        SyntaxKind.SimpleAssignmentExpression,
+                        IdentifierName(numElementsIdentifier),
+                        MemberAccessExpression(
+                            SyntaxKind.SimpleMemberAccessExpression,
+                            GetManagedValuesSource(info, context),
+                            IdentifierName("Length"))));
+                yield return GenerateByValueOutUnmarshalStatement(info, context);
             }
             else
             {
+                // <numElements> = <numElementsExpression>;
                 yield return ExpressionStatement(
                     AssignmentExpression(
                         SyntaxKind.SimpleAssignmentExpression,
@@ -682,50 +653,18 @@ namespace Microsoft.Interop
                                 Argument(IdentifierName(numElementsIdentifier))
                             })))));
 
-                string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-                string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
-
                 // ReadOnlySpan<TUnmanagedElement> <nativeSpan> = <marshallerType>.GetUnmanagedValuesSource(<nativeIdentifier>, <numElements>)
                 // Span<T> <managedSpan> = <marshallerType>.GetManagedValuesDestination(<managedIdentifier>)
                 // << unmarshal contents >>
-                yield return Block(
-                    LocalDeclarationStatement(VariableDeclaration(
-                        GenericName(
-                            Identifier(TypeNames.System_ReadOnlySpan),
-                            TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                        SingletonSeparatedList(
-                            VariableDeclarator(
-                                Identifier(nativeSpanIdentifier))
-                            .WithInitializer(EqualsValueClause(
-                                GetUnmanagedValuesSource(info, context)))))),
-                    LocalDeclarationStatement(VariableDeclaration(
-                        GenericName(
-                            Identifier(TypeNames.System_Span),
-                            TypeArgumentList(SingletonSeparatedList(_elementInfo.ManagedType.Syntax))),
-                        SingletonSeparatedList(
-                            VariableDeclarator(
-                                Identifier(managedSpanIdentifier))
-                            .WithInitializer(EqualsValueClause(
-                                GetManagedValuesDestination(info, context)))))),
-                    MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                        info,
-                        context,
-                        _elementInfo,
-                        _elementMarshaller,
-                        IdentifierName(numElementsIdentifier),
-                        StubCodeContext.Stage.UnmarshalCapture,
-                        StubCodeContext.Stage.Unmarshal));
+                yield return GenerateUnmarshalStatement(info, context);
             }
         }
 
         public bool UsesNativeIdentifier(TypePositionInfo info, StubCodeContext context) => true;
 
-        private static string GetNumElementsIdentifier(TypePositionInfo info, StubCodeContext context)
-            => context.GetAdditionalIdentifier(info, "numElements");
-
-        private InvocationExpressionSyntax GetUnmanagedValuesDestination(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetUnmanagedValuesDestination(TypePositionInfo info, StubCodeContext context)
         {
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
             string nativeIdentifier = context.GetIdentifiers(info).native;
 
             // <marshallerType>.GetUnmanagedValuesDestination(<nativeIdentifier>, <numElements>)
@@ -741,7 +680,7 @@ namespace Microsoft.Interop
                 })));
         }
 
-        private InvocationExpressionSyntax GetManagedValuesSource(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetManagedValuesSource(TypePositionInfo info, StubCodeContext context)
         {
             // <marshallerType>.GetManagedValuesSource(<managedIdentifer>)
             string managedIdentifier = context.GetIdentifiers(info).managed;
@@ -754,9 +693,9 @@ namespace Microsoft.Interop
                     Argument(IdentifierName(managedIdentifier)))));
         }
 
-        private InvocationExpressionSyntax GetUnmanagedValuesSource(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetUnmanagedValuesSource(TypePositionInfo info, StubCodeContext context)
         {
-            string numElementsIdentifier = GetNumElementsIdentifier(info, context);
+            string numElementsIdentifier = MarshallerHelpers.GetNumElementsIdentifier(info, context);
             string nativeIdentifier = context.GetIdentifiers(info).native;
 
             // <marshallerType>.GetUnmanagedValuesSource(<nativeIdentifier>, <numElements>)
@@ -772,7 +711,7 @@ namespace Microsoft.Interop
                 })));
         }
 
-        private InvocationExpressionSyntax GetManagedValuesDestination(TypePositionInfo info, StubCodeContext context)
+        protected override InvocationExpressionSyntax GetManagedValuesDestination(TypePositionInfo info, StubCodeContext context)
         {
             string managedIdentifier = context.GetIdentifiers(info).managed;
 
@@ -783,91 +722,6 @@ namespace Microsoft.Interop
                     _marshallerTypeSyntax,
                     IdentifierName(ShapeMemberNames.LinearCollection.Stateless.GetManagedValuesDestination)),
                 ArgumentList(SingletonSeparatedList(Argument(IdentifierName(managedIdentifier)))));
-        }
-
-        private StatementSyntax GenerateByValueUnmarshalStatement(TypePositionInfo info, StubCodeContext context)
-        {
-            // Use ManagedSource and NativeDestination spans for by-value marshalling since we're just marshalling back the contents,
-            // not the array itself.
-            // This code is ugly since we're now enforcing readonly safety with ReadOnlySpan for all other scenarios,
-            // but this is an uncommon case so we don't want to design the API around enabling just it.
-            string numElementsIdentifier = context.GetAdditionalIdentifier(info, "numElements");
-
-            // <numElements> = <marshallerType>.GetManagedValuesSource(<managedIdentifier>).Length;
-            ExpressionStatementSyntax numElementsStatement = ExpressionStatement(
-                AssignmentExpression(
-                    SyntaxKind.SimpleAssignmentExpression,
-                    IdentifierName(numElementsIdentifier),
-                    MemberAccessExpression(
-                        SyntaxKind.SimpleMemberAccessExpression,
-                        GetManagedValuesSource(info, context),
-                        IdentifierName("Length"))));
-
-            string managedSpanIdentifier = MarshallerHelpers.GetManagedSpanIdentifier(info, context);
-
-            // Span<TElement> <managedSpan> = MemoryMarshal.CreateSpan(ref Unsafe.AsRef(in <marshallerType>.GetManagedValuesSource(<managedIdentifier>).GetPinnableReference(), <numElements>));
-            LocalDeclarationStatementSyntax managedValuesDeclaration = LocalDeclarationStatement(VariableDeclaration(
-                GenericName(
-                    Identifier(TypeNames.System_Span),
-                    TypeArgumentList(
-                        SingletonSeparatedList(_elementInfo.ManagedType.Syntax))
-                ),
-                SingletonSeparatedList(VariableDeclarator(managedSpanIdentifier).WithInitializer(EqualsValueClause(
-                    InvocationExpression(
-                        MemberAccessExpression(
-                            SyntaxKind.SimpleMemberAccessExpression,
-                            ParseName(TypeNames.System_Runtime_InteropServices_MemoryMarshal),
-                            IdentifierName("CreateSpan")))
-                    .WithArgumentList(
-                        ArgumentList(
-                            SeparatedList(
-                                new[]
-                                {
-                                    Argument(
-                                        InvocationExpression(
-                                            MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
-                                                ParseName(TypeNames.System_Runtime_CompilerServices_Unsafe),
-                                                IdentifierName("AsRef")),
-                                            ArgumentList(SingletonSeparatedList(
-                                                Argument(
-                                                    InvocationExpression(
-                                                        MemberAccessExpression(
-                                                            SyntaxKind.SimpleMemberAccessExpression,
-                                                            GetManagedValuesSource(info, context),
-                                                            IdentifierName("GetPinnableReference")),
-                                                            ArgumentList()))
-                                                .WithRefKindKeyword(
-                                                    Token(SyntaxKind.InKeyword))))))
-                                    .WithRefKindKeyword(
-                                        Token(SyntaxKind.RefKeyword)),
-                                    Argument(
-                                        IdentifierName(numElementsIdentifier))
-                                }))))))));
-
-            // Span<TUnmanagedElement> <nativeSpan> = <marshallerType>.GetUnmanagedValuesDestination(<nativeIdentifier>, <numElements>)
-            string nativeSpanIdentifier = MarshallerHelpers.GetNativeSpanIdentifier(info, context);
-            LocalDeclarationStatementSyntax unmanagedValuesDeclaration = LocalDeclarationStatement(VariableDeclaration(
-                GenericName(
-                    Identifier(TypeNames.System_Span),
-                    TypeArgumentList(SingletonSeparatedList(_unmanagedElementType))),
-                SingletonSeparatedList(
-                    VariableDeclarator(
-                        Identifier(nativeSpanIdentifier))
-                    .WithInitializer(EqualsValueClause(
-                        GetUnmanagedValuesDestination(info, context))))));
-
-            return Block(
-                numElementsStatement,
-                managedValuesDeclaration,
-                unmanagedValuesDeclaration,
-                MarshallerHelpers.LinearCollection.NonBlittableContentsMarshallingStatement(
-                    info,
-                    context,
-                    _elementInfo,
-                    _elementMarshaller,
-                    IdentifierName(numElementsIdentifier),
-                    StubCodeContext.Stage.UnmarshalCapture,
-                    StubCodeContext.Stage.Unmarshal));
         }
     }
 }
